% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_toolbox.R
\name{run_toolbox}
\alias{run_toolbox}
\title{Run the CMSAF toolbox.}
\usage{
run_toolbox(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link[shiny]{runApp}}.}
}
\description{
Run the interactive shiny-based toolbox based on the cmsaf package.
}
\examples{
\dontrun{run_toolbox(launch.browser = TRUE)}
}
