% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmsaf.R
\docType{package}
\name{cmsaf}
\alias{cmsaf}
\title{cmsaf: A user-friendly Toolbox for preparing, manipulating, analyzing and visualizing 
satellite-based CM SAF NetCDF formatted data.}
\description{
The 'cmsafops' functions are manipulating NetCDF input files and write the result
in a separate output file. The functions were designed and tested for CM SAF
NetCDF data, but most of the functions can be applied to other NetCDF data,
which use the CF convention and time, latitude, longitude dimensions. As interface 
This package uses functionalities of the \link[cmsafops:cmsafops]{cmsafops} and
\link[cmsafvis:cmsafvis]{cmsafvis} packages.
}
\section{Get started}{
 Run \code{run_toolbox()} to open the interactive toolbox.
}

\references{
\url{http://www.cmsaf.eu/R_toolbox}

 Kothe, S.; Hollmann, R.; Pfeifroth, U.; Träger-Chatterjee, C.; Trentmann, J.
 The CM SAF R Toolbox—A Tool for the Easy Usage of Satellite-Based Climate
 Data in NetCDF Format. ISPRS Int. J. Geo-Inf. 2019, 8, 109.
 \doi{10.3390/ijgi8030109}
}
\author{
Maintainer: Steffen Kothe \email{Steffen.Kothe@dwd.de}

 Contact: CM SAF Team \email{contact.cmsaf@dwd.de}
}
\keyword{datagen}
\keyword{manip}
\keyword{package}
\keyword{spatial}
\keyword{ts}
\keyword{univar}
